Add files via upload
This is the Python code and report that I wrote for the Intro to Machine Learning course final project at the University of Texas at Austin. This was a graduate-level class taught in the Jackson School of Geoscience.

In my final project, I created a Bayesian Neural Network to model the relationship between two parameters which are often empirically related in the geotechnical engineering industry. These parameters include the standard penetration test blow count and the shear wave velocity of soil.